feat(quartzite-style): extract test block + add Container arm to DefaultStyle
Extract the 788-line test block from default_style.rs into a sibling
default_style_tests.rs (attached via #[path]), keeping the prod file at
~190 lines. Add the Container arm: fill_rect(Window) + draw_rect(WindowText
1px), no text, no child traversal.

2 new tests (AC1, AC2); all 49 tests green.
