+++ quartzite-style/Cargo.toml
@@ -20,6 +20,7 @@ nv-flip            = "0.1"
 image              = "0.25"
 tempfile           = "3"
 peniko             = "0.6"
+quartzite-core     = { path = "../quartzite-core" }
 quartzite-geometry = { path = "../quartzite-geometry" }
 quartzite-renderer = { path = "../quartzite-renderer" }
